+++ outbound-connectors/outbound-kafka/src/test/java/energy/eddie/outbound/kafka/KafkaTestConfig.java
@@ -4,6 +4,7 @@
 package energy.eddie.outbound.kafka;
 
 import energy.eddie.cim.agnostic.OpaqueEnvelope;
+import energy.eddie.cim.agnostic.PermissionCommand;
 import energy.eddie.cim.serde.MessageSerde;
 import energy.eddie.cim.serde.SerdeFactory;
 import energy.eddie.cim.serde.SerdeInitializationException;
@@ -118,6 +119,23 @@ public class KafkaTestConfig {
                                                                 new StringDeserializer(),
                                                                 new CustomDeserializer<>(serde, OpaqueEnvelope.class));
         var listenerContainerFactory = new ConcurrentKafkaListenerContainerFactory<String, OpaqueEnvelope>();
+        listenerContainerFactory.setConsumerFactory(consumerFactory);
+        return listenerContainerFactory;
+    }
+
+    @Bean
+    public KafkaListenerContainerFactory<ConcurrentMessageListenerContainer<String, PermissionCommand>> permissionCommandListenerContainerFactory(
+            @SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection") EmbeddedKafkaBroker embeddedKafka,
+            MessageSerde serde
+    ) {
+        var props = new HashMap<String, Object>();
+        props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, embeddedKafka.getBrokersAsString());
+        props.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest");
+        var consumerFactory = new DefaultKafkaConsumerFactory<>(props,
+                                                                new StringDeserializer(),
+                                                                new CustomDeserializer<>(serde,
+                                                                                         PermissionCommand.class));
+        var listenerContainerFactory = new ConcurrentKafkaListenerContainerFactory<String, PermissionCommand>();
         listenerContainerFactory.setConsumerFactory(consumerFactory);
         return listenerContainerFactory;
     }